requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via django
billiard==4.2.0
    # via celery
celery==5.4.0
    # via
    #   django-celery-beat
    #   django-celery-results
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.3
    # via django-celery-beat
diff-match-patch==20230430
    # via django-import-export
django==5.0.8
    # via
    #   django-celery-beat
    #   django-celery-results
    #   django-import-export
    #   django-jazzmin
    #   django-timezone-field
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-import-export==4.1.1
django-jazzmin==3.0.0
django-timezone-field==7.0
    # via django-celery-beat
kombu==5.4.0
    # via celery
pillow==10.4.0
prompt-toolkit==3.0.47
    # via click-repl
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   python-crontab
redis==5.0.8
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
tablib==3.5.0
    # via django-import-export
tzdata==2024.1
    # via
    #   celery
    #   django-celery-beat
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
